src/main/java/kitchenpos/menugroup/service/MenuGroupServiceJpa.java
package kitchenpos.menugroup.service;

import kitchenpos.menugroup.domain.MenuGroup;
import kitchenpos.menugroup.domain.MenuGroupRepository;
import kitchenpos.menugroup.dto.MenuGroupRequest;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Transactional
@Service
public class MenuGroupServiceJpa {
    private final MenuGroupRepository menuGroupRepository;

    public MenuGroupServiceJpa(MenuGroupRepository menuGroupRepository) {
        this.menuGroupRepository = menuGroupRepository;
    }

    public MenuGroup create(final MenuGroupRequest menuGroup) {
        return menuGroupRepository.save(menuGroup.of());
    }

    @Transactional(readOnly = true)
    public List<MenuGroup> list() {
        return menuGroupRepository.findAll();
    }
}
